requirements.txt
beautifulsoup4==4.5.3
blinker==1.3
chardet==2.3.0
click==6.7
#cloud-init==17.1
#command-not-found==0.3
configobj==5.0.6
#cryptography==1.2.3
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.12.2
#hibagent==1.0.1
idna==2.0
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
#language-selector==0.1
lxml==4.1.1
MarkupSafe==0.23
newspaper3k==0.2.5
nltk==3.2.5
numpy==1.13.3
oauthlib==1.0.3
Pillow==5.0.0
prettytable==0.7.2
pyasn1==0.1.9
pycurl==7.43.0
#pygobject==3.20.0
PyJWT==1.3.0
pyserial==3.0.1
#python-apt==1.1.0b1
python-dateutil==2.6.1
#python-debian==0.1.27
python-resize-image==1.1.11
#python-systemd==231
PyYAML==3.11
requests==2.13.0
requests-file==1.4.3
requests-oauthlib==0.8.0
six==1.10.0
#ssh-import-id==5.5
tldextract==2.2.0
tweepy==3.5.0
#ufw==0.35
#unattended-upgrades==0.1
urllib3==1.13.1
webpreview==1.3.2
Werkzeug==0.14.1
